requirements.txt
flask==2.0.1
flask-cors==3.0.10
opencv-python==4.7.0
pytesseract==0.3.10
numpy==1.24.3
Pillow==9.5.0
gunicorn==20.1.0
python-dotenv==0.19.0 